requirements.txt
numpy
matplotlib
scipy
pyyaml
inquirer